dt-bindings: soc: microchip: mpfs-mss-top-sysreg: add interrupt controller child
commit 14c7da9f39bcda392e9c71a3e58a125b67f9d447 from
https://github.com/linux4microchip/linux.git

On PolarFire SoC there are more GPIO interrupts than there are interrupt
lines available on the PLIC, and a runtime configurable mux is used to
decide which interrupts are assigned direct connections to the PLIC &
which are relegated to sharing a line.
This mux is, in our reference configuration, written by platform
firmware during boot based on the FPGA's configuration.

This interrupt controller is a child of the mss-top-sysreg simple-mfd.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>