media/platform/raspberypi/rp1_cfe: Candidate fix for #5821
commit 598462fb0279b3615c40430ce9f3616e0a9d44b8 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

To avoid lost frame start in a subsequent session, avoid setting
the number of lanes back to 1 or putting CSI-2 Host into reset.

It's not clear if this is a watertight fix -- what if the camera
itself produced a truncated or garbled packet, or continued to
send until the next start? -- but it does seem to fix the issue.

Signed-off-by: Nick Hollinghurst <nick.hollinghurst@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>