arm64: mpam: Add helpers to change a tasks and cpu mpam partid/pmg values
commit 791a50a6c1e0eff4014caccb85f7202bc9f0013d from
git@git.assembla.com:cavium/WindRiver.linux.git

Care must be taken when modifying the PARTID and PMG of a task, as
writing these values may race with the task being scheduled in, and
reading the modified values.

Add helpers to set the task properties, and the CPU default value.
These use WRITE_ONCE() that pairs with the READ_ONCE() in mpam_get_regval()
to avoid causing torn values.

CC: Dave Martin <Dave.Martin@arm.com>
CC: Ben Horgan <ben.horgan@arm.com>
Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Bo Sun <bo.sun.cn@windriver.com>