dts: 'AI Camera' uses releasing gated clock
commit  e6241fd7ca16b7b1ee35b577a5c9477fdb6f7c06 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

The clock using by IMX500 in 'AI Camera' is gpio-gated. The GPIO used is
provided by an I2C-controlled rpi-rp2040-gpio-bridge on 'AI Camera'.
Both the IMX500 and gpio-bridge share a common regulator for their power
supply. Using 'gpio-gate-clock', the 'AI Camera' will never be
powered-down as the GPIO provided by the gpio-bridge is always claimed
by the clock driver. Switching to 'gpio-gate-clock-releasing' causes the
GPIO to be released by the clock driver when the clock is not needed.
This allows 'AI Camera' to be powered-down.

Signed-off-by: Richard Oliver <richard.oliver@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>