media: i2c: imx500: GPIO acquire/release semantics
commit  d57bb1e4dd5ed50b92608d461275cde80a02b3f1 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

When the imx500 driver is used as part of the 'AI Camera', the poweroff
state is never reached as the camera and gpio driver share a regulator.
By releasing the GPIOs when they are not in use, 'AI Camera' is able to
achieve a powered-down state.

Signed-off-by: Richard Oliver <richard.oliver@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>