PENDING: media: ti: j721e-csi2rx: Serialize stream stops
commit e6c44d4f7164aed7c10d8b9136eb4841dfdf92a5 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Variable csi->enable_count tracks number of  active streams, it is
read to assert the pixel reset when stopping the last stream.

Currently the value of csi->enable_count is read without
synchronization, and it is decremented after calling disable_streams()
on the downstream element. This creates a race conditions when multiple
streams are stopped simultaneously, which could result in an unnecessary
pixel reset being asserted.

Serialize the read and decrement of csi->enable_count to prevent race
conditions.

Reviewed-by: Devarsh Thakkar <devarsht@ti.com>
Signed-off-by: Rishikesh Donadkar <r-donadkar@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>