PENDING: remoteproc: k3-r5: Return if core failed to suspend
commit 9ed014b81d215e14c929eed6a7ead9b3dd4613aa from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

If an R5 core fails to suspend, return the error. This allows for Linux
to know that the driver failed to suspend, rather than the remoteproc
driver having a false positive where it reports that all the R5 cores
suspended when they did not.

Fixes: 85225ba4f927 ("PENDING: remoteproc: k3-r5: add support for PM suspend/resume")
Signed-off-by: Kendall Willis <k-willis@ti.com>
Reviewed-by: Udit Kumar <u-kumar1@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>