LF-16045-3 arm64: dts: imx8mp: make 'dsp' node depend on 'aips5'
commit  c8dee768fb49f5bd8aceda0434efd9dbdc98a328 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

The DSP needs to access peripherals on AIPSTZ5 (to communicate with
the AP using AUDIOMIX MU, for instance). To do so, the security-related
registers of the bridge have to be configured before the DSP is started.
Enforce a dependency on AIPSTZ5 by adding the 'access-controllers'
property to the 'dsp' node.

Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>
Reviewed-by: Daniel Baluta <daniel.baluta@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>