LF-15161-2: Documentation: media: Describe exposure and gain controls for multiple captures
commit  f6529d42fa55a29cc3f7af5a1f0827632943a99a from
https://github.com/nxp-imx/linux-imx lf-6.18.y

The standard controls for exposure and gain allow a
single value for a single capture. For sensors with HDR
capabilities or context switching, this is not sufficient, so
new controls are added that allow multiple values—one for each
capture.

Signed-off-by: Mirela Rabulea <mirela.rabulea@nxp.com>
Reviewed-by: Laurentiu Palcu <laurentiu.palcu@oss.nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>