FROMLIST: arm64: dts: ti: k3-am69-sk: Add idle-states for remaining SERDES instances
commit 9337755c7f112c8052d25f03b36eb325bfe149f3 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

In AM69 SoC there are 4 instances of the 4 lane SERDES. So in
"serdes_ln_ctrl" node there are total 16 entries in "mux-reg-mask"
property. But "idle-states" is defined only for the lanes of first two
SERDES instances. For completeness, set the "idle-states" of lanes of
remaining SERDES instances to a default value of "unused".

Link: https://lore.kernel.org/all/20250609115921.2380611-1-h-salunke@ti.com
Signed-off-by: Hrushikesh Salunke <h-salunke@ti.com>
Reviewed-by: Siddharth Vadapalli <s-vadapalli@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>