v4l: xilinx: hdmirx: Add support for native interface mode
commit 94fc47bbcd2a0161b928527364f958bc40be9bb7 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

In native mode configuration, the HDMI IP does not use the
"s_axis_video_aclk" clock, and this clock is not present in the
device tree. The driver previously attempted to request this clock
unconditionally, causing probe to fail when the clock was absent in
native configurations. So skip the s_axis_video_aclk in native mode
configuration.

Signed-off-by: Rajesh Gugulothu <rajesh.gugulothu@amd.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
Message-ID: <1750161634-1209415-3-git-send-email-rajesh.gugulothu@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>