FROMLIST: drm/bridge: cdns-dsi: Drop checks that shouldn't be in .mode_valid()
commit fa5564c6eddbfa3c777d0c955c3398f202aac685 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The docs say about mode_valid():

"it is not allowed to look at anything else but the passed-in mode, and
validate it against configuration-invariant hardware constraints"

We're doing a lot more than just looking at the mode. The main issue
here is that we're doing checks based on the pixel clock, before we know
what the pixel clock from the crtc actually is.

So, drop the cdns_dsi_check_conf() call from .mode_valid().

Link: https://lore.kernel.org/all/20250618-cdns-dsi-impro-v4-9-862c841dbe02@ideasonboard.com/
Reviewed-by: Aradhya Bhatia <aradhya.bhatia@linux.dev>
Tested-by: Parth Pancholi <parth.pancholi@toradex.com>
Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
Signed-off-by: Harikrishna Shenoy <h-shenoy@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>