dt-bindings: clock: Add clock-output-names for Renesas ProXO
commit 3b994ebd5c761d1308f3761a0ab08be40f17dda8 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

clock-output-names is used by the driver but not described in binding
that's why add it. Also resort properties to have compatible as const first
and also remove it from required property because compatible is required
all the time.
At the same time also add it to renesas-soc.yaml which catch that it is not
approved compatible string. Likely this change is not correct but it is not
AMD ip to deal with that's why allow it to be able to check DTSes.

Signed-off-by: Michal Simek <michal.simek@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>