media: microchip-isc: Store histogram data of all channels
commit 9cc7f78e62b1f551e0677f1e20d572c8dbc377d0 from
https://github.com/linux4microchip/linux.git

Store the histogram data for all 4 channels(R, B, GR, GB). The data of each
channel is available on one interrupt. We need 4 interrupt cycles to get
all four channels. Store all the channels in a 2D array. This data is later
passed to the application buffer when it requests for a dequeue.

Signed-off-by: Balamanikandan Gunasundar <balamanikandan.gunasundar@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>