arch: arm: dts: Update calibration value for RTC nodes
commit 3286280c87142b69cfecb2988d858300f2f76d45 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

The Calibration Register is a 21-bit register and it comprises of a
16-bit Seconds Calibration Value, a 4-bit Fractional Calibration Value,
and a 1-bit Fractional Calibration Enable.

The 16-bit Seconds Calibration Value represents the number of Oscillator
Ticks that are required to measure the largest time period that is less
than or equal to 1 second.

For an oscillator that is 32.768 KHz, this value will be 0x7FFF.

Update calibration value to 0x7FFF in RTC nodes.

Signed-off-by: Harini T <harini.t@amd.com>
Message-ID: <20250627093118.218170-1-harini.t@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>