i2c: atr: Fix end of loop test in i2c_atr_find_mapping_by_addr()
commit 3d3e23a3479125e970fd493e64acde428f18057b from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

When the list_for_each_entry_reverse() exits without hitting a break
then the list cursor points to invalid memory.  So this check for
if (c2a->fixed) is checking bogus memory.  Fix it by using a "found"
variable to track if we found what we were looking for or not.

The list head (i2c_atr_chan.alias_pairs) is not a full entry, it's just
a struct list_head. When the for loop runs to completion, c2a doesn't
point to a struct i2c_atr_alias_pair, so you can't access c2a->fixed.

Fixes: c3f55241882b ("i2c: Support dynamic address translation")
Signed-off-by: Dan Carpenter <dan.carpenter@linaro.org>
Reviewed-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>
Reviewed-by: Romain Gantois <romain.gantois@bootlin.com>
Reviewed-by: Luca Ceresoli <luca.ceresoli@bootlin.com>
Signed-off-by: Wolfram Sang <wsa+renesas@sang-engineering.com>
State: upstream (20d5620e9b29)
Message-ID: <4d341b5602892fb1d4fac7248d0a32a250e37284.1751271753.git.michal.simek@amd.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>