i2c: atr: Fix lockdep for nested ATRs
commit 3a9b2900d5ac036f965231f332ae418e5cdf758c from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

When we have an ATR, and another ATR as a subdevice of the first ATR,
we get lockdep warnings for the i2c_atr.lock and
i2c_atr_chan.orig_addrs_lock. This is because lockdep uses a static key
for the locks, and doesn't see the locks of the separate ATR instances
as separate.

Fix this by generating a dynamic lock key per lock instance.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen+renesas@ideasonboard.com>
Signed-off-by: Cosmin Tanislav <demonsingur@gmail.com>
Reviewed-by: Luca Ceresoli <luca.ceresoli@bootlin.com>
Signed-off-by: Wolfram Sang <wsa+renesas@sang-engineering.com>
State: upstream (ec8c1b92eaec)
Message-ID: <c1ebcc8ad678bacc3730f1ca21d26b91c491617f.1751271753.git.michal.simek@amd.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>