i2c: atr: add static flag
commit 5dae4909ed7b840587ee271efcbded6a60750e06 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Some I2C ATRs do not support dynamic remapping, only static mapping
of direct children.

Mappings will only be added or removed as a result of devices being
added or removed from a child bus.

The ATR pool will have to be big enough to accommodate all devices
expected to be added to the child buses.

Add a new flag that prevents old mappings to be replaced or new mappings
to be created in the alias finding code paths. That mens adding a flags
parameter to i2c_atr_new() and an i2c_atr_flags enum.

Signed-off-by: Cosmin Tanislav <demonsingur@gmail.com>
Reviewed-by: Luca Ceresoli <luca.ceresoli@bootlin.com>
Reviewed-by: Romain Gantois <romain.gantois@bootlin.com>
Signed-off-by: Wolfram Sang <wsa+renesas@sang-engineering.com>
State: upstream (18355307dc56)
Message-ID: <daf221c5553cb277e64ba7ebbfcf81af52884b28.1751271753.git.michal.simek@amd.com>
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>