arm64: mpam: Advertise the CPUs MPAM limits to the driver
commit 8af57c04b8a6d9acc81ae995761595f9e678836c from
git@git.assembla.com:cavium/WindRiver.linux.git

Requestors need to populate the MPAM fields on the interconnect. For
the CPUs these fields are taken from the corresponding MPAMy_ELx
register. Each requestor may have a limit on the largest PARTID or
PMG value that can be used. The MPAM driver has to determine the
system-wide minimum supported PARTID and PMG values.

To do this, the driver needs to be told what each requestor's
limit is.

CPUs are special, but this infrastructure is also needed for the
SMMU and GIC ITS. Call the helper to tell the MPAM driver what the
CPUs can do.

The return value can be ignored by the arch code as it runs well
before the MPAM driver starts probing.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Bo Sun <bo.sun.cn@windriver.com>