net: enetc: Transition to phylink_fwnode_phy_connect()
commit  98765433692a1162bb142649ac6b0cb0da990a84 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

Break the dependency on the OF node for phylink connect
and create operations, by replacing the OF node with fwnode
references. Use a handy helper that prepares the transition
to swnode for the ENETC4 driver support for nodeless
pseudo-MACs, which will be introduced in a follow-up patch.

Signed-off-by: Claudiu Manoil <claudiu.manoil@nxp.com>
Reviewed-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>