staging: xilinx-tsn: add queue type information in vlan tci
commit 523da1431c670d7addfed06f482a0b8e2fb7e8d3 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

In IP designs with multiple TADMA queues, streams are assigned to
different queues, but identification relies solely on the MAC address
and VLAN ID. If a stream is set up for one TADMA queue, and traffic
with the same MAC and VLAN ID but a different PCP value (intended for
another queue) is sent, the driver will incorrectly route it to the
original queue, potentially disabling the sent queue. To resolve this,
include the queue type in the VLAN Tag Control Information.

Signed-off-by: Nagadheeraj Rottela <nagadheeraj.rottela@amd.com>
Reviewed-by: Harini Katakam <harini.katakam@amd.com>
Message-ID: <20250708112633.3701568-5-nagadheeraj.rottela@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>