LF-15707: mxc: vpu: wave6: Try to avoid additional dma sync
commit  64332543465d9e837f752f715a1ab0c7f849ffe2 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

When xen feature is eanbled, and if virtio-video is enabled, then the
dma-buffer may be allocated from swiotlb, not from cma, then the buffer
may be not coherent, to avoid data inconsistency, the driver performs a
force synchronization.

But in normal case, it may affect the performance, especially in the
case of multiple instances.

So if xen feature is not eabled, skip the force synchronization.
if driver is not built as module, check swiotlb pool before
synchronization.

Signed-off-by: Ming Qian <ming.qian@nxp.com>
Reviewed-by: Zhou Peng <eagle.zhou@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>