arm_mpam: resctrl: Pre-allocate free running monitors
commit c1b355c1a26c284ed759d78a741cb889c3308b2f from
git@git.assembla.com:cavium/WindRiver.linux.git

When there are enough monitors, the resctrl mbm local and total
files can be exposed. These need all the monitors that resctrl
may use to be allocated up front.

Add helpers to do this.

If a different candidate class is discovered, the old array
should be free'd and the allocated monitors returned to the
driver.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Bo Sun <bo.sun.cn@windriver.com>