ARM: dts: watchdog param creates early-watchdog
commit  9c09b75242960117155712f41ce540df2e3cd63c from
https://github.com/raspberrypi/linux.git rpi-6.12.y

The firmware change correcting the implementation of
dtoverlay_is_enabled had the unintended consequence of causing the
firmware to enable the watchdog even though the user had not explicitly
requested it. This is harmless on Linux because the watchdog driver
takes over and disarms it, but on other operating systems this can lead
to a reboot.

An upcoming change to the firmware will avoid this problem by also
requiring the presence of a new property, "early-watchdog". It's also
the case that disabling the watchdog driver in Device Tree also disables
important power management functionality. Therefore, change the
dtparam=watchdog implementation to only add the new property (or not),
leaving the container node enabled.

See: https://github.com/raspberrypi/firmware/issues/1980

Signed-off-by: Phil Elwell <phil@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>