firmware: tegra: bpmp: improve the solution of protecting critical code
When booting up preempt-rt kernel, there is below calltrace:
Call trace:
 ......
 __might_resched+0x12c/0x160
 rt_spin_lock+0x38/0xc8
 tegra_bpmp_transfer_atomic+0xa8/0x260
 tegra_bpmp_probe+0x1e8/0x3b0
 ......
 kernel_init_freeable+0x22c/0x2a0
 kernel_init+0x28/0x138
 ret_from_fork+0x10/0x20
Because in rt kernel, the spin_lock might enter sleep status and
it is called in atomic status, calltarce is reported. To fix this
issue, remove the irq disable/enable action when calling function
tegra_bpmp_transfer_atomic(), because it does not need to restrict
it to be called under atomic status. Accordingly, remove the code of
checking whether irq is in disabled status.

In addition, in order to protect critical code in both normal and
IRQ contexts, replace current spin_lock()/spin_unlock() with
spin_lock_irqsave()/spin_unlock_irqrestore()

Fixes: 1abb081e41a7 ("firmware: tegra: Simplify channel management")
Cc: stable@vger.kernel.org
Signed-off-by: Meng Li <Meng.Li@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>