tty: linflexuart: correct the dma_release_channel() place to fix the wrong suspending flow
In commit 3bf0bc2d1215 "tty: serial: linflexuart: UART support for FIFO with
DMA", the dma_release_channel() is wrongly put in the suspend execution path
when backporing the SDK patch, and this will cause suspend/resume failure.
So let's put it in the right place of the remove() function, then the failure
can be fixed.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>