dts: cm5: Fix Ethernet PHY interrupt pull-up
commit  40822060a09337c518194cd95bc7daf74ab244e2 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

On CM5, the active-low interrupt pin (INT_N) of the Ethernet PHY is
connected to GPIO37. However, an internal pull-up resistor appears to
be missing, which causes the interrupt edge to be missed or not detected
reliably. Fix this by configuring a bias pull-up on the gpio controller.

Signed-off-by: Nicolai Buchwitz <nb@tipi-net.de>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>