PENDING: drivers: crypto: sa2ul: Generate dynamic mdata length
commit d0f8500d5399cd603bcbcaef8a3d11317590ed3c from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Since the option bytes can now contain the AAD of different length, the
mdata length changes so we can't hardcode it anymore. Generate it
dynamically based on the content we send.

Also, while at it, remove the last 4 bytes that we were programming but
is not required and fix the offsets for the algorithms that got messed
up with it.

Also, the max size for the cmdl is 96 bytes so update the value accordingly

Signed-off-by: Manorit Chawdhry <m-chawdhry@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>