net: phy: Add phy shutdown function
There is such a situation that the network port will be
down during the execution of kexec -e. At this time, the
phy reset signal is always in the reset state. However,
when loading the kernel, the phy reset pin is always in
the reset state, which will cause failure in reading phy
id. Of course, it has no effect on some situations where
phy id is directly passed from DTB. It is difficult for
the current phy driver to have a better solution upstream,
so I added a phy shutdown as a workround to restore phy
reset-gpio to normal state during kexec -e.

Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>