clocksource/drivers/imx-sysctr: Initialize priv->cmpcr to 0
In commit 84abd535a3a6 ("LF-12085 clocksource/drivers/imx-sysctr:
fix i.MX95 support"), a problem was introduced. During the kdump
process, the initial value of the CMPCR register is not 0, but
when the timer is started, sysctr_set_state_shutdown() will be
called to disable the timer first. At this time, the CMPCR register
reads an abnormal value, and the CMPCR register will be cleared
after disbale timer, which will cause an infinite loop in
sysctr_timer_read_write() function. Currently, imx95 has not found
the reference manual, and the specific meaning of each bit in the
register cannot be checked for the time being. However, I detected
that the initial value of the system startup is 0, so it is reasonable
to set the default value to 0 as the workround.

Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>