arm_mpam: resctrl: Sort the order of the domain lists
commit 258751bf35c436a4312f8d0093d287379aff4b73 from
git@git.assembla.com:cavium/WindRiver.linux.git

resctrl documents that the domains appear in numeric order in the
schemata file. This means a little more work is needed when bringing
a domain online.

Add the support for this, using resctrl_find_domain() to find the
point to insert in the list.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Bo Sun <bo.sun.cn@windriver.com>