drm: rp1: Enable VEC->GPIO output; cosmetic change to registers
commit  83943264284590e75b95a0cecb3876924befe2f6 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

In the VEC driver, enable mapping VEC (not DPI) to DPI GPIOs.
This is to support VEC output over GPIO on Raspberry Pi CM5.
It is harmless as DPI and VEC could not be used concurrently,
and the output is anyway conditional on pinctrl.

Also, tweak the style of VIDEO_OUT_CFG register definitions
(in both DPI and VEC drivers) to be more Linux-friendly.

Signed-off-by: Nick Hollinghurst <nick.hollinghurst@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>