net: can: mpfs_can: avoid double write to RX_MSG#_CTRL_CMD registers
commit 4cd5c67afcc986dd184e5cf4da11810d18cea7ca from
https://github.com/linux4microchip/linux.git

The current code will write twice to the register, if the buffer is the
final once in a sequence. The first write is the same for all buffers,
and sets the "link flag" to tell the hardware that all buffers are in
the same sequence. This flag should not be set for the final buffer in a
sequence, as there are no following buffers to link it to, so the driver
writes again to the register to clear the bit it just set.

Rather than write twice, construct the value to be written as a
variable, setting the "link flag" as appropriate in the variable, and
write it just once.

Reported by a user doing Zephyr driver development.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>