net: can: mpfs_can: enable overload detection
commit 00fdbec4632f6117bd645a250967274016aeaa0f from
https://github.com/linux4microchip/linux.git

The driver contains code to handle overload frames, but never actually
enables them. There are no comments etc to explain why this is the case,
and so it appears to be an oversight. Enable the error's interrupt and
include it in the mask that wraps the code handling it.

Reported by a user doing Zephyr driver development.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>