LF-15623-6 net: enetc: extract netc_enable_cbdr() as a helper function
commit  5144a8a5eab7ea48964772ef74d7adef296ef094 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

When the switch or ENETC suspends, there is no need to free the command
BD ring's DMA buffer. This eliminates the need to re-allocate DMA memory
when resuming. Therefore, netc_enable_cbdr() is extracted from
netc_setup_cbdr() as a helper function to simplify the resume/suspend
process, and it will be used in a subsequent patch.

Another reason is that if switch or ENETC fails to resume, user space may
still have access to hardware resources. If the command BD ring's DMA
buffer is empty, accessing a null pointer could result in a crash.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>