octeontx2-af: Skip TM tree print for disabled SQs
commit beb286151c6e2ba93e07485cf09769ccf197dca5 from
git@git.assembla.com:cavium/WindRiver.linux.git

Currently, the TM tree is printing all SQ topology including those
which are not enabled, this results in redundant output for SQs
which are not active. This patch adds a check in print_tm_tree()
to skip printing the TM tree hierarchy if the SQ is not enabled.

Fixes: 95fbcdeed8c1 ("octeontx2-af: Add debugfs support to dump NIX TM registers")
Signed-off-by: Anshumali Gaur <agaur@marvell.com>
Change-Id: I9236f1f0b1e54d242b3ce31a28148593204d82c8
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/159138
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Bo Sun <bo.sun.cn@windriver.com>