LF-15961-1: arm64: dts: imx8mm-evk: reference the MUTE pinctrl node
commit  a9a1b080c73a43de693eb33526027c2cdfe93610 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

The pinctrl setting for 8mic MUTE is defined but not referenced, and
therefore not actually executed. Consequently, when the pin was
configured for other purposes, MUTE function would not work properly.
This patch references the MUTE pinctrl node to avoid this.

Signed-off-by: Lei Xu <lei.xu@nxp.com>
Reviewed-by: Chancel Liu <chancel.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>