LF-15961-2: arm64: dts: imx8mn-evk: reference the MUTE pinctrl node
commit  f27a637508438ebde8f73d36a6184cc4af360293 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

The pinctrl setting for 8mic MUTE is defined but not referenced, and
therefore not actually executed. Consequently, when the pin was
configured for other purposes, MUTE function would not work properly.
This patch references the MUTE pinctrl node to avoid this.

Signed-off-by: Lei Xu <lei.xu@nxp.com>
Reviewed-by: Chancel Liu <chancel.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>