LF-15961-3: arm64: dts: imx8mp: reference the MUTE pinctrl node
commit  3bace549991d1f9f6286d14b94856dc9d0a5ce34 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

The pinctrl setting for 8mic MUTE is defined but not referenced, and
therefore not actually executed. Consequently, when the pin was
configured for other purposes, MUTE function would not work properly.
This patch references the MUTE pinctrl node to avoid this.

Signed-off-by: Lei Xu <lei.xu@nxp.com>
Reviewed-by: Chancel Liu <chancel.liu@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>