overlays: Fix sc16is752-spi1 emulation
commit  14131859b2615fd3076ffdad370fc5500a037432 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

[1] Removed the sc16is752-spi1 overlay, replacing it with an entry in
overlay_map.dts that invokes sc16is75x-spi with specific parameters.
This does not work because it fails to configure the SPI1 interface.
Most such overlays would require the respective SPI interface to have
already been configured using one of the spi<n>-<m> overlays, but it is
not possible to do that using overlay_map, and it is unreasonable to
suddenly impose that requirement on users.

Work around that specific problem by adding an extra parameter to
sc16is75x to configure SPI1. It's not ideal, but better than a complete
dedicated overlay.

Link: https://github.com/raspberrypi/linux/issues/6962
Fixes: ce20a8fdbf58 ("overlays: sc16is75x: Add generic SPI overlay")
Signed-off-by: Phil Elwell <phil@raspberrypi.com>

[1] commit ce20a8fdbf58 ("overlays: sc16is75x: Add generic SPI overlay")
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>