PENDING: drm/bridge: it66121: Add it66122 support
commit 641579cfe4a4512e604272874d575c2d5b9acaa4 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The IT66122 is a drop in replacement for the IT66122. The part is
register compatible with what we use of the IT66121. The only relevant
change being the PID is now 0x0622 vs 0x0612. Add this extra PID
so probe does not fail during the PID check with these new parts.

Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>