drivers: media: pisp_be: Fix use after free in job queue logic
commit  cd18dc367f1b6b9985518410da7b8baa54ebee49 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

pispbe_schedule() currently takes a node group as a parameter, which is
left over from before the job prepare/scheduling refactoring. This is
now invalid, as jobs are executed in the order which they were queued.

As part of this old code, there was a check if the current node group
mached the node group of the job, and if unmathched, use a "continue"
statement. This is invalid as there is no loop to iterate over any more.
The reason this was not a compile bug is because of the for loop used
as part of the scoped_guard macro.

A consequence of breaking out of the scoped_guard loop early is that
the job structure gets freed, but not actually removed from the queue
and may be accessed after freeing.

Fix this by removing the node group test in pispbe_schedule() as it is
no longer valid to use.

Signed-off-by: Naushir Patuck <naush@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>