arm_mpam: resctrl: Reject oversized memory bandwidth portion bitmaps
commit a32dfc95fe61e22330a706af464ea1d1d3173954 from
git@git.assembla.com:cavium/WindRiver.linux.git

The MPAM architecture allows memory bandwidth portion bitmap (BWPBM)
controls to be up to 4096 bits in size.

Currently, the MPAM driver uses the scalar type u32 to represent the
bitmap internally, so only the first 32 partitions will be
allocatable.

This would render some of the hardware memory bandwidth unusable if
there are more partitions in the hardware, and there is no evidence yet
that any platform will need it.

Disable use of BWPBM controls that are larger than 32 bits, for now.

Signed-off-by: Dave Martin <Dave.Martin@arm.com>
Signed-off-by: Bo Sun <bo.sun.cn@windriver.com>