arm_mpam: resctrl: Fix MB min_bandwidth value exposed to userspace
commit ff47b7337ba1b4c01a15b0fdef85e07f2db750ff from
git@git.assembla.com:cavium/WindRiver.linux.git

Currently, the min_bandwidth value exposed to userspace in resctrlfs
for the MB resource does not match the way in which the value will get
converted during reading/writing of the schema.

It is also doubtful that the same rounding convention should be used
for bandwidth_gran as for min_bandwidth, since bandwidth_gran (as now
documented) is a bound on the worst-case precision of any value,
whereas min_bandwidth is the value that corresponds to the single,
specific minimum hardware encoding.

Introduce an explicit helper to generate the min value directly from
the conversion functions, and plumb it in.

Signed-off-by: Dave Martin <Dave.Martin@arm.com>
Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Bo Sun <bo.sun.cn@windriver.com>