gpio: mpfs: avoid overwriting parent chip data
commit 2ccbafa89796b4b77be97bba4199ee6db0cfa6a9 from
https://github.com/linux4microchip/linux.git

Unnoticed when parented by the irq mux, since it has no mask/unmask
functions, the driver was overwriting chip data for the parent interrupt
controller with a pointer to the gpio private date structure.
When interrupts are used with a coreGPIO, the parent controller is the
PLIC, which does have mask/unmask implementations. As a result, when the
PLIC driver attempts to convert the chip data to its private data
structure, it actually grabs the gpio structure and causes a kernel
crash. Stop setting chip data in the gpio driver and use a different
method to get the information into the interrupt handler.

Signed-off-by: Conor Dooley <conor.dooley@microchip.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>