PENDING: arm64: dts: ti: k3-am62l3-evm: use the correct padcfg macro
commit a6a3d865886d89f966d7d448765cb5710a067adc from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

While the padconfig layout has not changed for any of the K3 generation
of processors it can still cause confusion as to why the AM62PX_IOPAD()
macro would be used rather than the AM62LX_IOPAD() which was defined for
this SoC family.

So, to avoid any confusion, use the AM62LX_IOPAD() macro for every &pmx0
subnode for the AM62LX-EVM board.

Signed-off-by: Bryan Brattlof <bb@ti.com>
Reviewed-by: Judith Mendez <jm@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>