PENDING: net: ti: prueth: Adds Enhanced MultiCast filter support
commit 1bce32706917da518f795a303eeb3c777e831438 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

This patch addresses the issue with collisions in Multicast filter
in case of GOOSE and IEC92 traffic present on the link. Performing XOR
of a IEC92 MAC address will be identical to the XOR of a
GOOSE MAC address resulting in collision.

Extended the MC filter to hold extra bytes of information inorder to avoid
collision in the filter.

Reviewed-by: pmohan <pmohan@couthit.com>
Signed-off-by: Roger Quadros <rogerq@ti.com>
Signed-off-by: Andrew F. Davis <afd@ti.com>
Signed-off-by: Parvathi Pudi <parvathi@couthit.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>