dt-bindings: iio: adc: versal-sysmon: AI engine thermal monitoring
commit 96f828c080597ede91f501b527dc084ea7fdf692 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Add temperature child equivalent to supply child entries.
The sysmon temperature remote satellites can be exposed over this.

The reg property is the satellite number in the register database.
It is used for getting the register offset value. The xlnx,aie-temp
shows whether the temp child node is in close proximity with
AI engine. The name of the temperature satellite is added as the
xlnx,name property.

There is a limitation of reg collision between supply and temp
child nodes. Supply range is from 0 to 159 and temp range is from
1 to 64.

Signed-off-by: Anish Kadamathikuttiyil Karthikeyan Pillai <anish.kadamathikuttiyil-karthikeyan-pillai@amd.com>
Message-ID: <20250910122035.2046921-2-anish.kadamathikuttiyil-karthikeyan-pillai@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>