firmware: xilinx: Add support to clear TF-A PM state
commit 559626d3140b8f6c485043042fdb9918d52fc421 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Currently, during a kexec restart, only the kernel is reloaded, while the
TF-A specific data remain unchanged. This leads to a mismatch between the
kernel state and the TF-A firmware state like SGI number and shutdown
scope variable.

To resolve this, the TF_A_CLEAR_PM_STATE PM API is introduced to clear
TF-A-specific state during kexec. On a graceful reboot, this API is
triggered by zynqmp_firmware_shutdown(), while in a crash kernel scenario,
it is invoked by zynqmp_firmware_probe() in the reloaded kernel.

This support currently applies to Versal and Versal-Net platforms only.

Signed-off-by: Jay Buddhabhatti <jay.buddhabhatti@amd.com>
Acked-by: Senthil Nathan Thangaraj <senthilnathan.thangaraj@amd.com>
Message-ID: <20250912121950.1104451-2-jay.buddhabhatti@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>