FROMLIST: dt-bindings: serial: 8250_omap: Add wakeup pinctrl state
commit 51dc92b2e5cbdb56158abc0e8cb57d5242aa4625 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Pins associated with the 8250 omap unit can be the source of a wakeup in
deep sleep states. To be able to wakeup, these pins have to be
configured in a special way. To support this configuration add the
default and wakeup pinctrl states.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
Link: https://lore.kernel.org/all/20250910-uart-daisy-chain-8250-omap-v2-1-e90d44c1a9ac@ti.com/
Signed-off-by: Kendall Willis <k-willis@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>