drm/modes: Handle reflect_[xy] in the middle of the cmd line
commit  2c3d6946d702037eaf1bf4a47abf41541961f758 from
https://github.com/raspberrypi/linux.git rpi-6.12.y

The command line parser was looking for an "=" before the ","
separator.
If the command line had reflect_[xy] before another option
which took a parameter, it would find the "=" from the second
option and assume it was associated with the reflect option,
generally leading to a parsing failure.

Handle this case by looking for both "," and "=", and taking
the first instance.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>