drm/modes: Handle reflect_[xy] in the middle of the cmd line
commit  3e484f3404a99a05b44cfe719a6683b3e733940b from
https://github.com/raspberrypi/linux.git rpi-6.18.y

The command line parser was looking for an "=" before the ","
separator.
If the command line had reflect_[xy] before another option
which took a parameter, it would find the "=" from the second
option and assume it was associated with the reflect option,
generally leading to a parsing failure.

Handle this case by looking for both "," and "=", and taking
the first instance.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>