arm64: versal-net: update thermal sensor and node
commit d494108c3ef234ed9b715affb098b24c14d98efe from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Update the device tree to integrate the enhanced thermal
monitoring capability for the AI Engine.

A new thermal zone to monitor the AI Engine specifically is
added along with the existing thermal zone, with two trip info.

Also updated the io-channel-cells in sysmon0 node as 1
as per the dt binding.

Signed-off-by: Anish Kadamathikuttiyil Karthikeyan Pillai <anish.kadamathikuttiyil-karthikeyan-pillai@amd.com>
Message-ID: <20250923083019.2274145-5-anish.kadamathikuttiyil-karthikeyan-pillai@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>