phy: lynx-28g: implement phy_get_status() operations
commit  47d5f84adf15d90d9490f0db16de17391b603a22 from
https://github.com/nxp-imx/linux-imx lf-6.18.y

Allow consumers of the SerDes lane to figure out the PCS and AN/LT block
address without hardcoding it in the device tree or in the driver code,
which is what they do currently.

Also, implement CDR lock status, to gain a sense of when the link is
down for 25GbE where the PCS does not implement the power down
operation.

Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>