FROMLIST: media: chips-media: wave5: Fix Null reference while testing fluster
commit 40d7adcd233be37e8db917a9898144d5fe50930f from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

When multi instances are created/destroyed, many interrupts happens
and structures for decoder are removed.
"struct vpu_instance" this structure is shared for all flow in the decoder,
so if the structure is not protected by lock, Null dereference
could happens sometimes.
IRQ Handler was spilt to two phases and Lock was added as well.

Fixes: 9707a6254a8a ("media: chips-media: wave5: Add the v4l2 layer")
Link: https://patchwork.linuxtv.org/project/linux-media/patch/20250922055255.116-3-jackson.lee@chipsnmedia.com/
Signed-off-by: Jackson Lee <jackson.lee@chipsnmedia.com>
Signed-off-by: Nas Chung <nas.chung@chipsnmedia.com>
Signed-off-by: Brandon Brnich <b-brnich@ti.com>
Acked-by: Darren Etheridge <detheridge@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>