drm: xlnx: mmi-dc: Add video clock source selection property
commit 10f10e110a3486b663bb7481c35f21de6527b68b from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

The MMI DC supports switching between the PS video clock and PL video
clock via a glitch free mux. Add support for this switching by creating
a drm property and attach it to the MMI DC crtc. Both the pixel
clocks (if available) are programmed to the required mode clock so that
the switch to other video clock source happens instantly. The PL clock
is preferred to be used as pixel clock as it is more accurate than the
PS clock.

Reviewed-by: Anatoliy Klymenko <anatoliy.klymenko@amd.com>
Signed-off-by: Vishal Sagar <vishal.sagar@amd.com>
Message-ID: <20251006140708.2782229-2-vishal.sagar@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>