drm: xlnx: mmi-dc: Set MMI PLL rate to get accurate pixel clock
commit 0790afa81ad136aeb221366db0f7085525118abc from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

The MMI PLL is parent clock to aux0 and aux1 ref clocks. The aux0 and
aux1 ref clocks are required to be precise as they are the pixel clock and
audio video sync clock respectively.

If one of them is required to be changed, the MMI PLL currently doesn't
change, leading to inaccurate clocks getting generated.

Hence to get accurate clocks, the LCM of child clocks is calculated,
and the nearest value is set as parent MMI PLL clock to have least
deviation in the child clocks.

Signed-off-by: Rohit Visavalia <rohit.visavalia@amd.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
Reviewed-by: Anatoliy Klymenko <anatoliy.klymenko@amd.com>
Message-ID: <20251006140708.2782229-4-vishal.sagar@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>