LF-16411 net: enetc: add enetc_set_ringparam() to support changing rx-buf-len
commit  8b1f12c8d18e0ab7b4b6f61249a9d9881df7fbfb from
https://github.com/nxp-imx/linux-imx lf-6.18.y

The ENETC RX ring uses the page halves flipping mechanism, so the buffer
size of each RX BD is half a page. However, there is a problem as the
commit d1b15102dd16 ("net: enetc: add support for XDP_DROP and XDP_PASS")
mentioned.

"The only problem is that for default PAGE_SIZE values of 4096, buffer
sizes are 2048 bytes. While this is enough for normal skb allocations at
an MTU of 1500 bytes, for XDP it isn't, because the XDP headroom is 256
bytes, and including skb_shared_info and alignment, we end up being able
to make use of only 1472 bytes, which is insufficient for the default
MTU."

To solve the problem, the scatter/gather processing has been implemented
in the driver, so a packet of 1500 bytes is received in a buffer of 1472
bytes and another one of 28 bytes.

However, not all net devices support scatter/gather feature, some of
them due to hardware limitations and some of them have not implemented
the feature in the driver. Therefore, when ENETC receives a big frame
with multi-buffer and redirect it to other devices through XDP_REDIRECT
action, the frame will be dropped due to other devices do not support
scatter/gather feature.

Software limitations are easy to solve, but hardware limitations are not.
We can see two results for the devices with hardware limitations, either
the redirected packet is dropped if the packet received by multi-buffer,
or the device drivers reallocate a large buffer and copy data from these
non-linear buffers, which results in performance degradation or consumes
more CPU cycles. No matter what result is, the customers may not accept
it, especially those common packets in the network (length from 1473 to
1514 bytes excluding FCS).

Since the ethtool provides interface to change the Rx buffer size of the
RX ring, so we can leverage this interface to set the Rx buffer size as
users want. Currently, the supported rx-buf-len is 2^i * (PAGE_SIZE / 2),
the maximum is 32K.

Signed-off-by: Wei Fang <wei.fang@nxp.com>
Reviewed-by: Clark Wang <xiaoning.wang@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>