FROMLIST: dt-bindings: display: ti,am65x-dss: Add clk property for data edge synchronization
commit fedb651683740344cf882e3461dba1faa663a6a3 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

The dt-bindings for the display, specifically ti,am65x-dss, need to
include a clock property for data edge synchronization. The current
implementation does not correctly apply the data edge sampling property.

To address this, synchronization of writes to two different registers is
required: one in the TIDSS IP (which is already described in the tidss
node) and one is in the Memory Mapped Control Register Modules (added by
the previous commit).

As the Memory Mapped Control Register Modules is located in a different
IP, we need to use a phandle to write values in its registers.

Fixes: 32a1795f57ee ("drm/tidss: New driver for TI Keystone platform Display SubSystem")
Signed-off-by: Louis Chauvet <louis.chauvet@bootlin.com>
Link: https://lore.kernel.org/all/20250730-fix-edge-handling-v1-1-1bdfb3fe7922@bootlin.com/
Signed-off-by: Swamil Jain <s-jain1@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>