FROMLIST: drm/tidss: Fix sampling edge configuration
commit 9b02e9d3edd5cd6ab0c16df00b81f5a573284892 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

As stated in the AM62x Technical Reference Manual (SPRUIV7B), the data
sampling edge needs to be configured in two distinct registers: one in the
TIDSS IP and another in the memory-mapped control register modules. Since
the latter is not within the same address range, a phandle to a syscon
device is used to access the regmap.

Fixes: 32a1795f57ee ("drm/tidss: New driver for TI Keystone platform Display SubSystem")
Signed-off-by: Louis Chauvet <louis.chauvet@bootlin.com>
Link: http://lore.kernel.org/all/20250730-fix-edge-handling-v1-4-1bdfb3fe7922@bootlin.com/
Signed-off-by: Swamil Jain <s-jain1@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>