staging: mmi-dptx: Fix link up issue for some monitors
commit 9fb5bbfacf69c67a34b3d3483253e69f30e07cd7 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

Some monitors don't display the output when video data isn't sent in
some duration of time after link training to save power. Due to this,
the link training also gets reset. This leads to a blank screen when
video is sent.
Fix this issue by cycling the power state and doing link training with
the sink before enabling the video stream. The delays are added for the
display to settle and controller is configured correctly.

Signed-off-by: Vishal Sagar <vishal.sagar@amd.com>
Tested-by: Anatoliy Klymenko <anatoliy.klymenko@amd.com>
Reviewed-by: Anatoliy Klymenko <anatoliy.klymenko@amd.com>
Message-ID: <20251016021455.890184-1-vishal.sagar@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>