tty: linflexuart: add dma_tx_in_progress check to correct the dma tx bytes count
In linflexuart driver, when showing messages from kernel, like printk,
dev_info, etc. the linflex_console_write() will be called to put these
messages to console, and before this, linflex_stop_tx() will be invoked
to stop dma tx.

As linflex_stop_tx() is always called in atomic context, so it calls a async
function dmaengine_terminate_all() to terminate all active and pending dma
tx transfer to the related channel. Because the dmaengine_terminate_all() is
async, so it is possible the complete callback of a submitted transfer is still
running when this function returns, in this case, the complete callback
linflex_dma_tx_complete() may wrongly set the transferred bytes count, causing
kfifo out pointer was assigned an incorrect value, to overtake the in pointer,
which resulted in almost the entire contents of the tx buffer bing retransmitted
and loop messages printing.

So add a check for dma_tx_in_progress to correct the dma tx bytes count
to fix the above issue.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>