Revert "ARM: dts: Select the PL011 platform driver"
This patch adds the online serial port overlay function.
The kernel DTB will be passed to uboot through the bootloader.
This patch will cause the serial port function on uboot to be
abnormal. Uboot cannot output on the serial port, affecting the
use of uboot functions. At the same time, we currently do not
support the online DTB overlay function requirements, so revert it.

This reverts commit 9eb0c7efe0f419f01dcf50e0993048932d55cd8d.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>