boot: only emit rdinit warning on initramfs boot
commit 98aa4d5d242d3a73 [init/main.c: add warning when file specified in
rdinit is inaccessible] promoted a long time check to be visible on
boot.

The issue is that it is always issued even when an initramfs boot is
not used.

To avoid needing to completely disable CONFIG_BLK_DEV_INITRD and not
have the warning issues when an initramfs isn't used, we add checks for
the existence and size of an initramfs before allowing the warning
to be generated.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>