FROMLIST: clocksource/drivers/timer-ti-dm: Add clocksource support
commit c38062b74da6a8124d01ff8823c75040bf978aa5 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add support for using the TI Dual-Mode Timer as a clocksource. The
driver automatically picks the first timer that is marked as always-on
on with the "ti,timer-alwon" property to be the clocksource.

The timer can then be used for CPU independent time keeping.

Link: https://lore.kernel.org/all/20250623-topic-ti-dm-clkevt-v6-16-v1-0-b00086761ee1@baylibre.com/
Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>