FROMLIST: clocksource/drivers/timer-ti-dm: Add clockevent support
commit 4ae23ff8307957150d6dac39f6a5a1c8d14f420c from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Add support for using the TI Dual-Mode Timer for clockevents. The second
always on device with the "ti,timer-alwon" property is selected to be
used for clockevents. The first one is used as clocksource.

This allows clockevents to be setup independently of the CPU.

Link: https://lore.kernel.org/all/20250623-topic-ti-dm-clkevt-v6-16-v1-0-b00086761ee1@baylibre.com/
Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>