media: imx335: Update HBLANK range on mode change
commit  5aa823849baa128b8d4779ce9ed190afbc66efbb from
https://github.com/raspberrypi/linux.git rpi-6.18.y

commit d64ec84ac01b100619f68929a7e29f00617a6ffd upstream.

While switching modes, updating to a different value of HBLANK isn't
sufficient, as this is a read-only control with a single allowed value,
and thus hblank_min == hblank_max == hblank of the default mode.

So to correctly update the user-facing value of the HBLANK parameter,
which is necessary for correct framerate calculation, update the whole
range when switching modes.

Reviewed-by: Kieran Bingham <kieran.bingham@ideasonboard.com>
Signed-off-by: Jai Luthra <jai.luthra@ideasonboard.com>
Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Signed-off-by: Hans Verkuil <hverkuil+cisco@kernel.org>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>