drm: xlnx: mmi-dc: Clear CSC for disabled plane
commit de5735aeca8b8a2e8c94a9a1cf279571202439eb from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v6.12

MMI DC AV buffer is expected to produce black pixels when the corresponding
video layer source is set to 'None'. But it produces zero pixels instead,
which in turn are being treated by the blender according to the last active
pixel format. This results in a solid green screen after the YUV plane is
disabled. Fix this by resetting the blender's CSC configuration.

Signed-off-by: Anatoliy Klymenko <anatoliy.klymenko@amd.com>
Reviewed-by: Vishal Sagar <vishal.sagar@amd.com>
Tested-by: Vishal Sagar <vishal.sagar@amd.com>
Message-ID: <20251031191310.3385942-1-anatoliy.klymenko@amd.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>